+++ ripd/rip_snmp.c
@@ -317,7 +317,7 @@ rip2PeerLookup (struct variable *v, oid name[], size_t *length,
       peer = rip_peer_lookup (addr);
       if (peer)
 	{
-	  if ((len < sizeof (struct in_addr) + 1) ||
+	  if ((len < (int)sizeof (struct in_addr) + 1) ||
 	      (peer->domain > name[v->namelen + sizeof (struct in_addr)]))
 	    {
 	      oid_copy_addr (name + v->namelen, &peer->addr,
